+++ pkg/mapserver/trie/trie_test.go
@@ -353,19 +353,19 @@ func (d *MockDB) RetrieveUpdatedDomainHashesUpdates(ctx context.Context, perQuer
 
 func (d *MockDB) GetCountOfUpdatesDomainsUpdates(ctx context.Context) (int, error) { return 0, nil }
 
-func (d *MockDB) UpdateKeyValuesDomainEntries(ctx context.Context, keyValuePairs []db.KeyValuePair) (int, error) {
+func (d *MockDB) UpdateKeyValuesDomainEntries(ctx context.Context, keyValuePairs []db.KeyValuePair) (int64, error) {
 	return 0, nil
 }
 
-func (d *MockDB) UpdateKeyValuesTreeStruc(ctx context.Context, keyValuePairs []db.KeyValuePair) (int, error) {
+func (d *MockDB) UpdateKeyValuesTreeStruc(ctx context.Context, keyValuePairs []db.KeyValuePair) (int64, error) {
 	return 0, nil
 }
 
-func (d *MockDB) DeleteKeyValuesTreeStruc(ctx context.Context, keys []common.SHA256Output) (int, error) {
+func (d *MockDB) DeleteKeyValuesTreeStruc(ctx context.Context, keys []common.SHA256Output) (int64, error) {
 	return 0, nil
 }
 
-func (d *MockDB) AddUpdatedDomainHashesUpdates(ctx context.Context, keys []common.SHA256Output) (int, error) {
+func (d *MockDB) AddUpdatedDomainHashesUpdates(ctx context.Context, keys []common.SHA256Output) (int64, error) {
 	return 0, nil
 }
